Version estable factura ,tiquete credito

+++ src/main/java/com/emprendesoftcr/Bo/AbonoBo.java
@@ -12,7 +12,7 @@ public interface AbonoBo {
 
 	void eliminar(Abono abono);
 
-	Abono buscar(Integer id);
+	Abono buscar(Long id);
 
 	Collection<Abono> buscarPorCuentaCobrar();
