+++ src/main/java/com/emprendesoftcr/Bo/MotivoEntradaBo.java
@@ -11,7 +11,7 @@ public interface MotivoEntradaBo {
 
 	void eliminar(MotivoEntrada motivoEntrada);
 
-	MotivoEntrada buscar(Integer id);
+	MotivoEntrada buscar(Long id);
 
 	MotivoEntrada buscarPorDescripcionYEmpresa(String descripcion, Empresa empresa);
