+++ src/main/java/com/emprendesoftcr/Bo/ProveedorBo.java
@@ -11,7 +11,7 @@ public interface ProveedorBo {
 
 	void eliminar(Proveedor proveedor);
 
-	Proveedor buscar(Integer id);
+	Proveedor buscar(Long id);
 
 	Proveedor buscarPorNombreCompletoYEmpresa(String nombreCompleto, Empresa empresa);
