+++ src/main/java/com/emprendesoftcr/Bo/VendedorBo.java
@@ -12,7 +12,7 @@ public interface VendedorBo {
 
 	void eliminar(Vendedor vendedor);
 
-	Vendedor buscar(Integer id);
+	Vendedor buscar(Long id);
 
 	Vendedor buscarPorNombreCompletoYEmpresa(String nombreCompleto, Empresa empresa);
