+++ src/main/java/com/emprendesoftcr/Dao/ArticuloDao.java
@@ -11,7 +11,7 @@ public interface ArticuloDao {
 
 	void eliminar(Articulo articulo);
 
-	Articulo buscar(Integer id);
+	Articulo buscar(Long id);
 
 	Articulo buscarPorDescripcionYEmpresa(String descripcion, Empresa empresa);
