+++ src/main/java/com/emprendesoftcr/Dao/ClienteDao.java
@@ -11,7 +11,7 @@ public interface ClienteDao {
 
 	void eliminar(Cliente cliente);
 
-	Cliente buscar(Integer id);
+	Cliente buscar(Long id);
 
 	Cliente buscarPorNombreCompletoYEmpresa(String nombreCompleto, Empresa empresa);
