+++ src/main/java/com/emprendesoftcr/Dao/CompraDao.java
@@ -11,7 +11,7 @@ public interface CompraDao {
 
 	void eliminar(Compra compra);
 
-	Compra findById(Integer id);
+	Compra findById(Long id);
 
 	Compra findByConsecutivoAndEmpresa(String consecutivo, Empresa empresa);
