+++ src/main/java/com/emprendesoftcr/Dao/CuentaCobrarDao.java
@@ -11,7 +11,7 @@ public interface CuentaCobrarDao {
 
 	void eliminar(CuentaCobrar cuentaCobrar);
 
-	CuentaCobrar buscar(Integer id);
+	CuentaCobrar buscar(Long id);
 	CuentaCobrar buscarPorLetraCambio(String letraCambio);
 	CuentaCobrar buscarPorFacturaManual(Integer facturaManual);
