+++ src/main/java/com/emprendesoftcr/Dao/FacturaDao.java
@@ -11,7 +11,7 @@ public interface FacturaDao {
 
 	void eliminar(Factura factura);
 
-	Factura findById(Integer id);
+	Factura findById(Long id);
 
 	Factura findByConsecutivoAndEmpresa(String consecutivo, Empresa empresa);
