+++ src/main/java/com/emprendesoftcr/modelo/Categoria.java
@@ -32,7 +32,7 @@ public class Categoria implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id")
-	private Integer						id;
+	private Long						id;
 
 	@Column(name = "descripcion")
 	private String						descripcion;
@@ -62,7 +62,9 @@ public class Categoria implements Serializable {
 
 	}
 
-	public Categoria(Integer id, String descripcion, String estado, Date created_at, Date updated_at, Empresa empresa) {
+
+
+	public Categoria(Long id, String descripcion, String estado, Date created_at, Date updated_at, Empresa empresa) {
 		super();
 		this.id = id;
 		this.descripcion = descripcion;
@@ -72,13 +74,23 @@ public class Categoria implements Serializable {
 		this.empresa = empresa;
 	}
 
-	public Integer getId() {
+
+
+
+
+	
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+
+
+	
+	public void setId(Long id) {
 		this.id = id;
 	}
+
+
 
 	public String getDescripcion() {
 		return descripcion;
